release/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0
minio>=7.1.0
httpx>=0.24.0
python-dotenv>=1.0.0
email-validator>=2.0.0
pydantic-settings>=2.0.0
cryptography>=41.0.0